+++ src/app/app.component.ts
@@ -128,7 +128,7 @@ export class AppComponent implements OnInit {
           filter((userAppVersionDetails) => !!userAppVersionDetails)
         )
         .subscribe((userAppVersionDetails) => {
-          const { appSupportDetails, lastLoggedInVersion, eou, deviceInfo } = userAppVersionDetails;
+          const { appSupportDetails, lastLoggedInVersion, deviceInfo } = userAppVersionDetails;
           this.trackingService.eventTrack('Auto Logged out', {
             lastLoggedInVersion,
             appVersion: deviceInfo.appVersion,
