+++ src/app/core/models/add-attachment-properties.model.ts
@@ -1,0 +1,5 @@
+export interface AddAttachmentProperties {
+  type: string;
+  Mode: string;
+  Category: string;
+}
