+++ src/app/core/models/app-launch-time-properties.model.ts
@@ -1,0 +1,5 @@
+export interface AppLaunchTimeProperties {
+  'App launch time': string;
+  'Is logged in': boolean;
+  'Is multi org': boolean;
+}
