+++ src/app/core/models/capture-single-receipt-time-properties.model.ts
@@ -1,0 +1,5 @@
+export interface CaptureSingleReceiptTimeProperties {
+  'Capture receipt time': string;
+  'Is logged in': boolean;
+  'Is multi org': boolean;
+}
