+++ src/app/core/models/card-enrolled-properties.model.ts
@@ -1,0 +1,7 @@
+import { CardNetworkType } from '../enums/card-network-type';
+
+export interface CardEnrolledProperties {
+  Source: string;
+  'Card Network': CardNetworkType;
+  'Card ID': string;
+}
