+++ src/app/core/models/card-enrollment-errors-properties.model.ts
@@ -1,0 +1,7 @@
+import { CardNetworkType } from '../enums/card-network-type';
+
+export interface CardEnrollmentErrorsProperties {
+  'Card Network': CardNetworkType;
+  Source: string;
+  'Error Message': string;
+}
