+++ src/app/core/models/card-unenrolled-properties.model.ts
@@ -1,0 +1,6 @@
+import { CardNetworkType } from '../enums/card-network-type';
+
+export interface CardUnenrolledProperties {
+  'Card Network': CardNetworkType;
+  'Card ID': string;
+}
