+++ src/app/core/models/comment-history-action-properties.model.ts
@@ -1,0 +1,4 @@
+export interface CommentHistoryActionProperties {
+  action: string;
+  segment: string;
+}
