+++ src/app/core/models/corporate-card-expense-properties.model.ts
@@ -1,0 +1,6 @@
+import { Transaction } from './v1/transaction.model';
+
+export interface CorporateCardExpenseProperties {
+  Type: string;
+  transaction: Transaction;
+}
