+++ src/app/core/models/create-report-properties.model.ts
@@ -1,0 +1,4 @@
+export interface CreateReportProperties {
+  Expense_Count: number;
+  Report_Value: number;
+}
