+++ src/app/core/models/enrolling-non-rtf-card-properties.model.ts
@@ -1,0 +1,3 @@
+export interface EnrollingNonRTFCardProperties {
+  Source: string;
+}
