+++ src/app/core/models/event-track-properties.model.ts
@@ -1,0 +1,11 @@
+import { DeviceInfo } from '@capacitor/device';
+
+export interface EventTrackProperties {
+  Asset: string;
+  DeviceType: string;
+  deviceInfo: Partial<DeviceInfo>;
+  appVersion: string;
+  label: string;
+  lastLoggedInVersion: string;
+  user_email: string;
+}
