+++ src/app/core/models/expense-click-properties.model.ts
@@ -1,0 +1,6 @@
+import { ExpenseView } from './expense-view.enum';
+
+export interface ExpenseClickProperties {
+  view: ExpenseView;
+  category: string;
+}
