+++ src/app/core/models/filter-pill-clicked-properties.model.ts
@@ -1,0 +1,4 @@
+export interface FilterPillClickedProperties {
+  Asset: string;
+  filterPillType: string;
+}
