+++ src/app/core/models/footer-button-click-properties.model.ts
@@ -1,0 +1,4 @@
+export interface FooterButtonClickProperties {
+  Action: string;
+  Url: string;
+}
