+++ src/app/core/models/identify-properties.model.ts
@@ -1,0 +1,3 @@
+export interface IdentifyProperties {
+  [key: string]: string;
+}
