+++ src/app/core/models/on-setting-toggle-properties.model.ts
@@ -1,0 +1,5 @@
+export interface OnSettingToggleProperties {
+  userSetting: string;
+  action: string;
+  setDefaultCurrency: boolean;
+}
