+++ src/app/core/models/policy-correction-properties.model.ts
@@ -1,0 +1,4 @@
+export interface PolicyCorrectionProperties {
+  Violation: string;
+  Mode: string;
+}
