+++ src/app/core/models/report-name-change-properties.model.ts
@@ -1,0 +1,4 @@
+export interface ReportNameChangeProperties {
+  Time_spent: number;
+  Roles: string[];
+}
