+++ src/app/core/models/switch-Oorg-properties.model.ts
@@ -1,0 +1,9 @@
+export interface SwitchOrgProperties {
+  Asset: string;
+  'Switch To': string;
+  'Is Destination Org Active': boolean;
+  'Is Destination Org Primary': boolean;
+  'Is Current Org Primary': boolean;
+  Source: string;
+  'User Org ID': string;
+}
