+++ src/app/core/models/switch-org-launch-time-properties.model.ts
@@ -1,0 +1,4 @@
+export interface SwitchOrgLaunchTimeProperties {
+  'Switch org launch time': string;
+  'Login method': string;
+}
