+++ src/app/core/models/task-filter-clear-all-properties.model.ts
@@ -1,0 +1,6 @@
+import { TaskFilters } from './task-filters.model';
+
+export interface TaskFilterClearAllProperties {
+  Asset: string;
+  appliedFilters: TaskFilters;
+}
