+++ src/app/core/models/task-page-open-properties.model.ts
@@ -1,0 +1,4 @@
+export interface TaskPageOpenProperties {
+  Asset: string;
+  from: string;
+}
