+++ src/app/core/models/task-properties.model.ts
@@ -1,0 +1,4 @@
+export interface TaskProperties {
+  Asset: string;
+  header: string;
+}
