+++ src/app/core/models/view-report-info-properties.model.ts
@@ -1,0 +1,7 @@
+import { ExpenseView } from './expense-view.enum';
+
+export interface ViewReportInfoProperties {
+  view: ExpenseView;
+  action: string;
+  segment: string;
+}
